move processing into parse_config.py
