+++ py3status/parse_config.py
@@ -4,7 +4,38 @@ import glob
 import os
 import imp
 from collections import OrderedDict
+from string import Template
 
+
+MAX_NESTING_LEVELS = 4
+
+TIME_FORMAT = '%Y-%m-%d %H:%M:%S'
+TZTIME_FORMAT = '%Y-%m-%d %H:%M:%S %Z'
+TIME_MODULES = ['time', 'tztime']
+
+I3STATUS_MODULES = [
+    'battery', 'cpu_temperature', 'disk', 'ethernet', 'path_exists',
+    'run_watch', 'tztime', 'volume', 'wireless'
+]
+I3S_SINGLE_NAMES = ['cpu_usage', 'ddate', 'ipv6', 'load', 'time']
+ERROR_CONFIG = '''
+general {colors = true interval = 60}
+
+order += "static_string py3status"
+order += "tztime local"
+order += "group error"
+
+static_string py3status {format = "py3status"}
+tztime local {format = "%c"}
+group error{
+    button_next = 1
+    button_prev = 0
+    fixed_width = False
+    format = "{output}"
+    static_string error_min {format = "CONFIG ERROR" color = "#FF0000"}
+    static_string error {format = "$error" color = "#FF0000"}
+}
+'''
 
 class ParseException(Exception):
     def __init__(self, error, line, line_no, position, token):
@@ -21,7 +52,8 @@ class ParseException(Exception):
     def __str__(self):
         marker = ' ' * (self.position - 1) + '^'
         return '{}\n\nsaw `{}` at line {} position {}\n\n{}\n{}'.format(
-            self.error, self.token, self.line_no, self.position, self.line, marker)
+            self.error, self.token, self.line_no, self.position, self.line,
+            marker)
 
 
 class ModuleDefinition(OrderedDict):
@@ -64,14 +96,7 @@ class ConfigParser:
         module_path = os.path.join(root, 'modules', '*.py')
         for file in glob.glob(module_path):
             modules.append(os.path.basename(file)[:-3])
-        i3status_modules = [
-            'battery', 'cpu_temperature', 'disk',
-            'ethernet', 'path_exists', 'run_watch',
-            'tztime', 'volume', 'wireless'
-        ]
-        i3s_single_names = ['cpu_usage', 'ddate', 'ipv6', 'load', 'time']
-        self.module_names = modules + i3s_single_names + i3status_modules
-        self.i3s_single_names = i3s_single_names
+        self.module_names = modules + I3S_SINGLE_NAMES + I3STATUS_MODULES
         self.container_modules = []
 
     def check_child_friendly(self, name):
@@ -101,7 +126,7 @@ class ConfigParser:
         if split_name[0] not in self.module_names:
             self.current_token -= len(split_name) - offset
             self.error('Unknown module')
-        if len(split_name) > 1 and split_name[0] in self.i3s_single_names:
+        if len(split_name) > 1 and split_name[0] in I3S_SINGLE_NAMES:
             self.current_token -= len(split_name) - 1 - offset
             self.error('Invalid name cannot have 2 tokens')
         if len(split_name) > 2:
@@ -271,7 +296,7 @@ class ConfigParser:
                 if t_value == end_token:
                     raise self.ParseEnd()
                 elif t_value == '\n':
-                        continue
+                    continue
             if token['type'] == 'literal':
                 return self.make_value(t_value)
             elif t_value == '[':
@@ -281,15 +306,15 @@ class ConfigParser:
             elif t_value == '(':
                 return tuple(self.make_list(end_token=')'))
             else:
-                self.error('Value expected', previous=not(end_token))
+                self.error('Value expected', previous=not (end_token))
 
     def module_def(self):
         '''
         This is a module definition so parse content till end.
         '''
-        self.module_level += 1
-        if self.module_level == 3:
+        if self.module_level == MAX_NESTING_LEVELS:
             self.error('Module nested too deep')
+        self.module_level += 1
         module = ModuleDefinition()
         self.parse(module, end_token='}')
         self.module_level -= 1
@@ -342,7 +367,7 @@ class ConfigParser:
                     self.error('Name expected')
                 elif t_value == '+=' and name not in dictionary:
                     # order is treated specially
-                    if not(self.level == 1 and name == 'order'):
+                    if not (self.level == 1 and name == 'order'):
                         self.error('{} does not exist'.format(name))
                 if t_value in ['{']:
                     if self.current_module:
@@ -371,6 +396,141 @@ class ConfigParser:
         '''
         pass
 
+
+def process_config(config_path, py3_wrapper):
+    """
+    Parse i3status.conf so we can adapt our code to the i3status config.
+    """
+    general_defaults = {
+        'color_bad': '#FF0000',
+        'color_degraded': '#FFFF00',
+        'color_good': '#00FF00',
+        'color_separator': '#333333',
+        'colors': False,
+        'interval': 5,
+        'output_format': 'i3bar',
+    }
+
+    config = {}
+
+    user_modules = py3_wrapper.get_user_modules()
+    with open(config_path, 'r') as f:
+        try:
+            config_info = parse_config(f, user_modules=user_modules)
+        except ParseException as e:
+
+            error = e.one_line()
+            py3_wrapper.notify_user(error)
+            error_config = Template(ERROR_CONFIG).substitute(error=error.replace('"', '\\"'))
+            config_info = parse_config(error_config)
+
+    # update general section with defaults
+    if 'general' in config_info:
+        general_defaults.update(config_info['general'])
+    config['general'] = general_defaults
+
+    # get all modules
+    modules = {}
+    on_click = {}
+    i3s_modules = set()
+    py3_modules = set()
+    module_groups = {}
+    group_extras = []
+
+    def process_onclick(key, value, group_name):
+        try:
+            button = int(key.split()[1])
+            if button not in range(1, 6):
+                raise ValueError('should be 1, 2, 3, 4 or 5')
+        except IndexError as e:
+            raise IndexError(
+                'missing "button id" for "on_click" '
+                'parameter in group {}'.format(group_name))
+        except ValueError as e:
+            raise ValueError('invalid "button id" '
+                             'for "on_click" parameter '
+                             'in group {} ({})'.format(
+                                 group_name, e))
+        clicks = on_click.setdefault(group_name, {})
+        clicks[button] = value
+
+    def get_module_type(name):
+        if name.split()[0] in I3S_SINGLE_NAMES + I3STATUS_MODULES:
+            return 'i3status'
+        return 'py3status'
+
+    def process_module(name, module, parent):
+        if parent:
+            modules[parent]['items'].append(name)
+            mg = module_groups.setdefault(name, [])
+            mg.append(parent)
+            group_extras.append(name)
+            if get_module_type(name) == 'py3status':
+                module['.group'] = parent
+
+        # check module content
+        for k, v in module.items():
+            if k.startswith('on_click'):
+                # on_click event
+                process_onclick(k, v, name)
+            if isinstance(v, ModuleDefinition):
+                # we are a container
+                module['items'] = []
+        return module
+
+    def get_modules(data, parent=None):
+        for k, v in data.items():
+            if isinstance(v, ModuleDefinition):
+                module = process_module(k, v, parent)
+                modules[k] = module
+                get_modules(v, parent=k)
+
+    get_modules(config_info)
+
+    config['order'] = []
+    config['.group_extras'] = []
+
+    def fix_module(module):
+        fixed = {}
+        for k, v in module.items():
+            if not isinstance(v, ModuleDefinition):
+                fixed[k] = v
+        return fixed
+
+    def update_config(name, order):
+        module = modules.get(name, {})
+        module_type = get_module_type(name)
+        if order:
+            config['order'].append(name)
+        elif module_type == 'i3status':
+            config['.group_extras'].append(name)
+        if module_type == 'i3status':
+            i3s_modules.add(name)
+        else:
+            py3_modules.add(name)
+        config[name] = fix_module(module)
+
+    for name in config_info['order']:
+        update_config(name, order=True)
+
+    for name in group_extras:
+        update_config(name, order=False)
+
+    config['on_click'] = on_click
+    config['i3s_modules'] = sorted(list(i3s_modules))
+    config['py3_modules'] = sorted(list(py3_modules))
+    config['.module_groups'] = module_groups
+
+    # time and tztime modules need a format for correct processing
+    for name in config:
+        if name.split()[0] in TIME_MODULES and 'format' not in config[
+                name]:
+            if name.split()[0] == 'time':
+                config[name]['format'] = TIME_FORMAT
+            else:
+                config[name]['format'] = TZTIME_FORMAT
+
+    return config
 
 def parse_config(config, user_modules=None):
     '''
